requirements.txt
python-dotenv==0.12.0
python-telegram-bot==11.1.0
redis==3.4.1
requests==2.23.0
vk-api==11.8.0
